feat: add struct for game objects
struct groups texture, object and common methods

refact(game class): change game logic

Game uses struct of core and render types

reduce size of Game and more readable, clean code

refact(coordinate): change coordinate to use double

using speed * time for moving

feat(physics): add physics, collision detection, move methods

add speed concepts, collision detection for entity and maze

refact(bot): change bot movement logic

bot moves using speed * time with path direction to player

refact(textures): texture class serve as base

feat(camera): add struct camera in render

separate implementation, increase reusability

+++ core/src/Coordinates.cpp
@@ -1,6 +1,9 @@
-#include "coordinate.h"
+#include "Coordinates.h"
 
-Coordinate::Coordinate(int32_t x, int32_t y)
+/* -------------------------------------------------- */
+// constructors
+/* -------------------------------------------------- */
+Coordinate::Coordinate(double x, double y)
 	: x(x),
 	  y(y) 
 { }
@@ -15,6 +18,10 @@ Coordinate::Coordinate(Coordinate&& c) noexcept
 	  y(c.y) 
 { }
 
+
+/* -------------------------------------------------- */
+// =, ==, != operators with rhl and lhl values
+/* -------------------------------------------------- */
 bool Coordinate::operator == (const Coordinate& c) const 
 	{ return (c.x == this->x && c.y == this->y); }
